+++ xchange-coinbase/src/main/java/org/knowm/xchange/coinbase/v2/CoinbaseAuthenticated.java
@@ -20,7 +20,6 @@ import org.knowm.xchange.coinbase.v2.dto.account.CoinbasePaymentMethodsData;
 import org.knowm.xchange.coinbase.v2.dto.account.CoinbaseSellData;
 import org.knowm.xchange.coinbase.v2.dto.account.CoinbaseTransactionsResponse;
 import org.knowm.xchange.coinbase.v2.dto.account.transactions.CoinbaseBuySellResponse;
-import org.knowm.xchange.coinbase.v3.CoinbaseV3Digest;
 import si.mazi.rescu.ParamsDigest;
 
 @Path("/v2")
@@ -32,8 +31,8 @@ public interface CoinbaseAuthenticated extends Coinbase {
    *
    * <p>All request bodies should have content type application/json and be valid JSON.
    *
-   * <p>The CB-ACCESS-SIGN header is generated by creating a sha256 HMAC using the secret key on the
-   * prehash string timestamp + method + requestPath + body (where + represents string
+   * <p>The CB-ACCESS-SIGN header is generated by creating a sha256 HMAC using the secret key on
+   * the prehash string timestamp + method + requestPath + body (where + represents string
    * concatenation). The timestamp value is the same as the CB-ACCESS-TIMESTAMP header.
    *
    * <p>The body is the request body string or omitted if there is no request body (typically for
@@ -54,120 +53,89 @@ public interface CoinbaseAuthenticated extends Coinbase {
 
   @GET
   @Path("accounts/{accountId}/transactions")
-  CoinbaseTransactionsResponse getTransactions(
-      @HeaderParam(CB_VERSION) String apiVersion,
+  CoinbaseTransactionsResponse getTransactions(@HeaderParam(CB_VERSION) String apiVersion,
       @HeaderParam(CB_ACCESS_KEY) String apiKey,
       @HeaderParam(CB_ACCESS_SIGN) ParamsDigest signature,
       @HeaderParam(CB_ACCESS_TIMESTAMP) BigDecimal timestamp,
-      @PathParam("accountId") String accountId)
-      throws IOException, CoinbaseException;
+      @PathParam("accountId") String accountId) throws IOException, CoinbaseException;
 
   @GET
   @Path("accounts/{accountId}/buys")
-  CoinbaseBuySellResponse getBuys(
-      @HeaderParam(CB_VERSION) String apiVersion,
+  CoinbaseBuySellResponse getBuys(@HeaderParam(CB_VERSION) String apiVersion,
       @HeaderParam(CB_ACCESS_KEY) String apiKey,
       @HeaderParam(CB_ACCESS_SIGN) CoinbaseV2Digest signature,
       @HeaderParam(CB_ACCESS_TIMESTAMP) BigDecimal timestamp,
-      @PathParam("accountId") String accountId,
-      @QueryParam("limit") Integer limit,
-      @QueryParam("starting_after") String startingAfter)
-      throws IOException, CoinbaseException;
+      @PathParam("accountId") String accountId, @QueryParam("limit") Integer limit,
+      @QueryParam("starting_after") String startingAfter) throws IOException, CoinbaseException;
 
   @GET
   @Path("accounts/{accountId}/sells")
-  CoinbaseBuySellResponse getSells(
-      @HeaderParam(CB_VERSION) String apiVersion,
+  CoinbaseBuySellResponse getSells(@HeaderParam(CB_VERSION) String apiVersion,
       @HeaderParam(CB_ACCESS_KEY) String apiKey,
       @HeaderParam(CB_ACCESS_SIGN) CoinbaseV2Digest signature,
       @HeaderParam(CB_ACCESS_TIMESTAMP) BigDecimal timestamp,
-      @PathParam("accountId") String accountId,
-      @QueryParam("limit") Integer limit,
-      @QueryParam("starting_after") String startingAfter)
-      throws IOException, CoinbaseException;
+      @PathParam("accountId") String accountId, @QueryParam("limit") Integer limit,
+      @QueryParam("starting_after") String startingAfter) throws IOException, CoinbaseException;
 
   @GET
   @Path("accounts/{accountId}/deposits")
-  Map getDeposits(
-      @HeaderParam(CB_VERSION) String apiVersion,
+  Map getDeposits(@HeaderParam(CB_VERSION) String apiVersion,
       @HeaderParam(CB_ACCESS_KEY) String apiKey,
       @HeaderParam(CB_ACCESS_SIGN) ParamsDigest signature,
       @HeaderParam(CB_ACCESS_TIMESTAMP) BigDecimal timestamp,
-      @PathParam("accountId") String accountId)
-      throws IOException, CoinbaseException;
+      @PathParam("accountId") String accountId) throws IOException, CoinbaseException;
 
   @GET
   @Path("accounts/{accountId}/withdrawals")
-  Map getWithdrawals(
-      @HeaderParam(CB_VERSION) String apiVersion,
+  Map getWithdrawals(@HeaderParam(CB_VERSION) String apiVersion,
       @HeaderParam(CB_ACCESS_KEY) String apiKey,
       @HeaderParam(CB_ACCESS_SIGN) ParamsDigest signature,
       @HeaderParam(CB_ACCESS_TIMESTAMP) BigDecimal timestamp,
-      @PathParam("accountId") String accountId)
-      throws IOException, CoinbaseException;
+      @PathParam("accountId") String accountId) throws IOException, CoinbaseException;
 
   @GET
   @Path("accounts")
   CoinbaseAccountsData getAccounts(
       @HeaderParam(CB_AUTHORIZATION_KEY) ParamsDigest authTokenGenerator,
-      @QueryParam("limit") Integer limit,
-      @QueryParam("starting_after") String starting_after)
+      @QueryParam("limit") Integer limit, @QueryParam("starting_after") String starting_after)
       throws IOException, CoinbaseException;
 
   @GET
   @Path("accounts/{currency}")
-  CoinbaseAccountData getAccount(
-      @HeaderParam(CB_VERSION) String apiVersion,
-      @HeaderParam(CB_ACCESS_KEY) String apiKey,
-      @HeaderParam(CB_ACCESS_SIGN) CoinbaseV2Digest signature,
-      @HeaderParam(CB_ACCESS_TIMESTAMP) BigDecimal timestamp,
-      @PathParam("currency") String currency)
-      throws IOException, CoinbaseException;
+  CoinbaseAccountData getAccount(@HeaderParam(CB_AUTHORIZATION_KEY) ParamsDigest authTokenGenerator,
+      @PathParam("currency") String currency) throws IOException, CoinbaseException;
 
   @POST
   @Path("accounts")
   @Consumes(MediaType.APPLICATION_JSON)
-  CoinbaseAccountData createAccount(
-      @HeaderParam(CONTENT_TYPE) String contentType,
-      @HeaderParam(CB_VERSION) String apiVersion,
-      @HeaderParam(CB_ACCESS_KEY) String apiKey,
+  CoinbaseAccountData createAccount(@HeaderParam(CONTENT_TYPE) String contentType,
+      @HeaderParam(CB_VERSION) String apiVersion, @HeaderParam(CB_ACCESS_KEY) String apiKey,
       @HeaderParam(CB_ACCESS_SIGN) String signature,
-      @HeaderParam(CB_ACCESS_TIMESTAMP) BigDecimal timestamp,
-      Object payload)
+      @HeaderParam(CB_ACCESS_TIMESTAMP) BigDecimal timestamp, Object payload)
       throws IOException, CoinbaseException;
 
   @GET
   @Path("payment-methods")
-  CoinbasePaymentMethodsData getPaymentMethods(
-      @HeaderParam(CB_VERSION) String apiVersion,
+  CoinbasePaymentMethodsData getPaymentMethods(@HeaderParam(CB_VERSION) String apiVersion,
       @HeaderParam(CB_ACCESS_KEY) String apiKey,
       @HeaderParam(CB_ACCESS_SIGN) CoinbaseV2Digest signature,
-      @HeaderParam(CB_ACCESS_TIMESTAMP) BigDecimal timestamp)
-      throws IOException, CoinbaseException;
+      @HeaderParam(CB_ACCESS_TIMESTAMP) BigDecimal timestamp) throws IOException, CoinbaseException;
 
   @POST
   @Path("accounts/{account}/buys")
   @Consumes(MediaType.APPLICATION_JSON)
-  CoinbaseBuyData buy(
-      @HeaderParam(CONTENT_TYPE) String contentType,
-      @HeaderParam(CB_VERSION) String apiVersion,
-      @HeaderParam(CB_ACCESS_KEY) String apiKey,
+  CoinbaseBuyData buy(@HeaderParam(CONTENT_TYPE) String contentType,
+      @HeaderParam(CB_VERSION) String apiVersion, @HeaderParam(CB_ACCESS_KEY) String apiKey,
       @HeaderParam(CB_ACCESS_SIGN) String signature,
       @HeaderParam(CB_ACCESS_TIMESTAMP) BigDecimal timestamp,
-      @PathParam("account") String accountId,
-      Object payload)
-      throws IOException, CoinbaseException;
+      @PathParam("account") String accountId, Object payload) throws IOException, CoinbaseException;
 
   @POST
   @Path("accounts/{account}/sells")
   @Consumes(MediaType.APPLICATION_JSON)
-  CoinbaseSellData sell(
-      @HeaderParam(CONTENT_TYPE) String contentType,
-      @HeaderParam(CB_VERSION) String apiVersion,
-      @HeaderParam(CB_ACCESS_KEY) String apiKey,
+  CoinbaseSellData sell(@HeaderParam(CONTENT_TYPE) String contentType,
+      @HeaderParam(CB_VERSION) String apiVersion, @HeaderParam(CB_ACCESS_KEY) String apiKey,
       @HeaderParam(CB_ACCESS_SIGN) String signature,
       @HeaderParam(CB_ACCESS_TIMESTAMP) BigDecimal timestamp,
-      @PathParam("account") String accountId,
-      Object payload)
-      throws IOException, CoinbaseException;
+      @PathParam("account") String accountId, Object payload) throws IOException, CoinbaseException;
 }
