--- xchange-coinbase/src/main/java/org/knowm/xchange/coinbase/v3/dto/orders/CoinbaseOrdersResponse.java
@@ -1,5 +1,0 @@
-package org.knowm.xchange.coinbase.v3.dto.orders;
-
-public class CoinbaseOrdersResponse {
-
-}
