--- xchange-coinbase/src/main/java/org/knowm/xchange/coinbase/v3/dto/pricebook/CoinbasePriceBooksResponse.java
@@ -1,5 +1,0 @@
-package org.knowm.xchange.coinbase.v3.dto.pricebook;
-
-public class CoinbasePriceBooksResponse {
-
-}
