--- xchange-coinbase/src/main/java/org/knowm/xchange/coinbase/v3/dto/products/CoinbaseProductCandlesResponse.java
@@ -1,5 +1,0 @@
-package org.knowm.xchange.coinbase.v3.dto.products;
-
-public class CoinbaseProductCandlesResponse {
-
-}
